Fixed melee attack on tiles when effects are present (window cracks)
